feat: migration to new kubeadm-based cluster
Unfortunately, we can't get OpenEBS to come along for the ride, no
matter how hard I try, so I'll just manually update this config with
specific PVs on an as-needed basis.

I'm having a hell of a time getting PostgreSQL to run inside K8s anyway,
so I think I'll just move to a dedicated single SQL database hosted on
the non-K8s machine, anyway, so I can stop fighting that battle as well.
At that point, there should basically not be any need for PVs.